+++ zetaclient/chains/solana/signer/execute.go
@@ -6,7 +6,7 @@ import (
 
 	"cosmossdk.io/errors"
 	"github.com/ethereum/go-ethereum/common"
-	"github.com/gagliardetto/solana-go"
+	sol "github.com/gagliardetto/solana-go"
 	"github.com/near/borsh-go"
 	"github.com/rs/zerolog"
 
@@ -81,10 +81,10 @@ func (signer *Signer) createMsgExecute(
 		return nil, nil, errors.Wrap(err, "cannot validate sender")
 	}
 
-	remainingAccounts := []*solana.AccountMeta{}
+	remainingAccounts := []*sol.AccountMeta{}
 	for _, a := range msg.Accounts {
-		remainingAccounts = append(remainingAccounts, &solana.AccountMeta{
-			PublicKey:  solana.PublicKey(a.PublicKey),
+		remainingAccounts = append(remainingAccounts, &sol.AccountMeta{
+			PublicKey:  sol.PublicKey(a.PublicKey),
 			IsWritable: a.IsWritable,
 		})
 	}
@@ -105,7 +105,7 @@ func (signer *Signer) createMsgExecute(
 }
 
 // createExecuteInstruction wraps the execute 'msg' into a Solana instruction.
-func (signer *Signer) createExecuteInstruction(msg contracts.MsgExecute) (*solana.GenericInstruction, error) {
+func (signer *Signer) createExecuteInstruction(msg contracts.MsgExecute) (*sol.GenericInstruction, error) {
 	// create execute instruction with program call data
 	discriminator := contracts.DiscriminatorExecute
 	var dataBytes []byte
@@ -114,7 +114,7 @@ func (signer *Signer) createExecuteInstruction(msg contracts.MsgExecute) (*solan
 		serializedInst, err := borsh.Serialize(contracts.ExecuteRevertInstructionParams{
 			Discriminator: discriminator,
 			Amount:        msg.Amount(),
-			Sender:        solana.MustPublicKeyFromBase58(msg.Sender()),
+			Sender:        sol.MustPublicKeyFromBase58(msg.Sender()),
 			Data:          msg.Data(),
 			Signature:     msg.SigRS(),
 			RecoveryID:    msg.SigV(),
@@ -149,15 +149,15 @@ func (signer *Signer) createExecuteInstruction(msg contracts.MsgExecute) (*solan
 		return nil, errors.Wrap(err, "cannot decode connected pda address")
 	}
 
-	predefinedAccounts := []*solana.AccountMeta{
-		solana.Meta(signer.relayerKey.PublicKey()).WRITE().SIGNER(),
-		solana.Meta(signer.pda).WRITE(),
-		solana.Meta(msg.To()).WRITE(),
-		solana.Meta(destinationProgramPda).WRITE(),
+	predefinedAccounts := []*sol.AccountMeta{
+		sol.Meta(signer.relayerKey.PublicKey()).WRITE().SIGNER(),
+		sol.Meta(signer.pda).WRITE(),
+		sol.Meta(msg.To()).WRITE(),
+		sol.Meta(destinationProgramPda).WRITE(),
 	}
 	allAccounts := append(predefinedAccounts, msg.RemainingAccounts()...)
 
-	inst := &solana.GenericInstruction{
+	inst := &sol.GenericInstruction{
 		ProgID:        signer.gatewayID,
 		DataBytes:     dataBytes,
 		AccountValues: allAccounts,
@@ -178,7 +178,7 @@ func validateSender(sender string, executeType contracts.ExecuteType) (string, e
 	}
 
 	// for revert execute, sender should be a Solana address
-	senderSol, err := solana.PublicKeyFromBase58(sender)
+	senderSol, err := sol.PublicKeyFromBase58(sender)
 	if err != nil {
 		return "", errors.Wrapf(err, "invalid execute revert sender %s", sender)
 	}