+++ changelog.md
@@ -24,6 +24,7 @@
 * [4205](https://github.com/zeta-chain/node/pull/4205) - remove index field in ballot
 * [4200](https://github.com/zeta-chain/node/pull/4200) - remove `LastBlockHeight` state variable
 * [4174](https://github.com/zeta-chain/node/pull/4174) - add documentation for ZetaClient logging fields
+* [4213](https://github.com/zeta-chain/node/pull/4213) - prepare the client interfaces of the observer-signers for dry mode
 
 ### Fixes